Added meetup categories in selection tag format and combined meetup details with meetup class
