+++ src/main/java/launchcode/liftoff_project/Model/MeetupCategory.java
@@ -1,5 +1,6 @@
 //package launchcode.liftoff_project.Model;
 //
+//
 //import javax.persistence.Entity;
 //import javax.persistence.OneToMany;
 //import javax.validation.constraints.Size;
